Refactor cleanFile function to handle missing file name

+++ src/utils/cleanFile.ts
@@ -3,6 +3,10 @@ import { log } from "./logger"
 
 export default async function cleanFile(fileName: string) {
 	try {
+		if (!fileName)
+			return log.warn(
+				"A file was not cleaned because it was not provided"
+			)
 		await fs.promises.unlink(fileName)
 		log.info("A file was cleaned: ", fileName)
 	} catch (err) {
